REQ: Ajustes valida calculo valueTotalEndoso

+++ L5 Functions/Gosocket.Dian.Plugin.Functions/Common/LogicalEventRadian.cs
@@ -2221,7 +2221,7 @@ namespace Gosocket.Dian.Plugin.Functions.Common
                 }
 
                 //Calculo valor de la negociación
-                double resultNegotiationValue = (Int32.Parse(valueTotalEndoso, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture) * (100 - Int32.Parse(valueDiscountRateEndoso, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture)));
+                double resultNegotiationValue = (double.Parse(valueTotalEndoso, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture) * (100 - double.Parse(valueDiscountRateEndoso, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture)));
                 resultNegotiationValue = resultNegotiationValue / 100;
 
                 //Se debe comparar el valor de negociación contra el saldo(Nuevo Valor en disponibilización)
